refactor(economy): module structure cleanup
