+++ src/modules/economy/components/DepositForm/index.ts
@@ -1,0 +1,1 @@
+export * from './CreateDepositForm'
