+++ src/modules/economy/components/index.ts
@@ -1,0 +1,5 @@
+export * from './DepositForm'
+export * from './AccountCard'
+export * from './MyDeposits'
+export * from './MyExpenditures'
+export * from './MyPurchases'
