--- src/modules/economy/index.ts
@@ -1,6 +1,0 @@
-export * from './Deposits'
-export * from './mutations'
-export * from './MyEconomy'
-export * from './queries'
-export * from './types'
-export * from './utils'
