+++ src/modules/economy/mutations.hooks.ts
@@ -1,6 +1,6 @@
 import { useMutation } from '@apollo/client'
 import { CREATE_DEPOSIT_MUTATION } from './mutations'
-import { CreateDepositMutationReturns } from './types'
+import { CreateDepositMutationReturns } from './types.graphql'
 
 export function useDepositMutations() {
   const [createDeposit, { loading: createDepositLoading }] =
