+++ src/modules/economy/types.graphql.ts
