+++ src/modules/economy/views/index.ts
@@ -1,1 +1,3 @@
 export * from './CreateDeposit'
+export * from './Deposits'
+export * from './MyEconomy'
